Apply refactored valueItem classes
