+++ src/main/java/bisq/monitor/server/handlers/DaoStateHandler.java
@@ -18,10 +18,10 @@
 package bisq.monitor.server.handlers;
 
 import bisq.common.util.Tuple2;
-import bisq.core.monitor.ReportingItems;
 import bisq.monitor.reporter.Metrics;
 import bisq.monitor.reporter.Reporter;
 import bisq.monitor.utils.Util;
+import bisq.seednode.reporting.ReportingItems;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.*;
