+++ src/main/java/bisq/monitor/server/handlers/NetworkDataHandler.java
@@ -17,8 +17,8 @@
 
 package bisq.monitor.server.handlers;
 
-import bisq.core.monitor.ReportingItems;
 import bisq.monitor.reporter.Reporter;
+import bisq.seednode.reporting.ReportingItems;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
